ai-services/transaction-sage/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-dotenv
httpx
SQLAlchemy
PyJWT
python-jose[cryptography]